requirements.txt
fastapi==0.115.12
uvicorn==0.34.1
langchain==0.3.20
langgraph==0.3.5
langgraph-checkpoint-postgres==2.0.15
psycopg2-binary==2.9.10
python-dotenv==1.0.0
torch==2.2.1
torchvision==0.17.1
torchaudio==2.2.1
transformers==4.37.2
tokenizers==0.15.2
google-cloud-storage>=2.15.0
azure-storage-blob>=12.19.0
Pillow>=10.2.0
langchain-chroma>=0.0.1
langchain-openai>=0.0.1


